Move google strategy out of routes
